pin the gross-up inverse and move the pool limit validation out
